backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
pydantic-settings
python-dotenv
python-multipart